hud: move around/repackage model and view
