+++ internal/hud/hudmodel/model_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/windmilleng/tilt/internal/hud/view"
+	"github.com/windmilleng/tilt/internal/hud/hudview"
 	"github.com/windmilleng/tilt/internal/model"
 )
 
@@ -14,21 +14,21 @@ const manifestName = model.ManifestName("vigoda")
 func TestNewViewNormal(t *testing.T) {
 	vtf := newViewTestFixture(t)
 	vtf.run()
-	vtf.assertStatus(view.ResourceStatusFresh, "Running")
+	vtf.assertStatus(hudview.ResourceStatusFresh, "Running")
 }
 
 func TestNewViewNoPod(t *testing.T) {
 	vtf := newViewTestFixture(t)
 	delete(vtf.model.Pods, manifestName)
 	vtf.run()
-	vtf.assertStatus(view.ResourceStatusStale, "No pod found")
+	vtf.assertStatus(hudview.ResourceStatusStale, "No pod found")
 }
 
 func TestNewViewBrokenPod(t *testing.T) {
 	vtf := newViewTestFixture(t)
 	vtf.model.Pods[manifestName].Status = "CrashLoopBackOff"
 	vtf.run()
-	vtf.assertStatus(view.ResourceStatusBroken, "CrashLoopBackOff")
+	vtf.assertStatus(hudview.ResourceStatusBroken, "CrashLoopBackOff")
 }
 
 type viewTestFixture struct {
@@ -36,7 +36,7 @@ type viewTestFixture struct {
 	resource *Resource
 	model    *upperState
 
-	generatedView view.View
+	generatedView hudview.View
 }
 
 func newViewTestFixture(t *testing.T) *viewTestFixture {
@@ -61,7 +61,7 @@ func newViewTestFixture(t *testing.T) *viewTestFixture {
 			},
 		},
 	}
-	return &viewTestFixture{t, &r, &m, view.View{}}
+	return &viewTestFixture{t, &r, &m, hudview.View{}}
 }
 
 func (vtf *viewTestFixture) run() {
@@ -77,7 +77,7 @@ func (vtf *viewTestFixture) run() {
 	assert.Equal(vtf.t, vtf.resource.DirectoryWatched, ret.Resources[0].DirectoryWatched)
 }
 
-func (vtf *viewTestFixture) assertStatus(expectedStatus view.ResourceStatus, desc string) {
+func (vtf *viewTestFixture) assertStatus(expectedStatus hudview.ResourceStatus, desc string) {
 	assert.Equal(vtf.t, expectedStatus, vtf.generatedView.Resources[0].Status)
 	assert.Equal(vtf.t, desc, vtf.generatedView.Resources[0].StatusDesc)
 }
