+++ internal/hud/renderer.go
@@ -1,11 +1,11 @@
 package hud
 
-import "github.com/windmilleng/tilt/internal/hud/view"
+import "github.com/windmilleng/tilt/internal/hud/hudview"
 
 type Renderer struct {
 }
 
-func (r Renderer) Render(v view.View) error {
+func (r Renderer) Render(v hudview.View) error {
 	// TODO: draw v
 	return nil
 }
